+++ src/tokenizer/token.rs
@@ -1,10 +1,10 @@
 use crate::location::Location;
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub enum Token {
     String(String),
     Identifier(String),
-    Number(u32),
+    Number(f64),
 
     OpenBrace,
     CloseBrace,
